TaskShared, EMPTYDBS, utils.taskDebug, utils.assertWarning, improved setNestedProperties, Jest in tests, DB namespace

+++ processor/rxjs/src/storage.mjs
@@ -4,7 +4,7 @@ License, v. 2.0. If a copy of the MPL was not distributed with this
 file, You can obtain one at https://mozilla.org/MPL/2.0/.
 */
 
-import { COPROCESSOR, MONGO_URL } from "../config.mjs";
+import { COPROCESSOR, MONGO_URL, EMPTYDBS } from "../config.mjs";
 import * as dotenv from "dotenv";
 dotenv.config();
 import mongoose from 'mongoose';
@@ -25,20 +25,22 @@ db.once('open', () => {
   console.log("Connected to MongoDB!");
 });
 
-db.once('open', async () => {
-  try {
-      const collections = Object.keys(mongoose.connection.collections);
-      for (const collectionName of collections) {
-          const collection = mongoose.connection.collections[collectionName];
-          await collection.drop();
-          console.log(`Dropped ${collectionName} collection successfully!`);
-      }
-  } catch (err) {
-      if (err.message !== 'ns not found') { // Ignore if the collection is not found
-          console.error('Error dropping collection:', err);
-      }
-  }
-});
+if (EMPTYDBS) {
+  db.once('open', async () => {
+    try {
+        const collections = Object.keys(mongoose.connection.collections);
+        for (const collectionName of collections) {
+            const collection = mongoose.connection.collections[collectionName];
+            await collection.drop();
+            console.log(`Dropped ${collectionName} collection successfully!`);
+        }
+    } catch (err) {
+        if (err.message !== 'ns not found') { // Ignore if the collection is not found
+            console.error('Error dropping collection:', err);
+        }
+    }
+  });
+}
 
 const coprocessorPrefix = COPROCESSOR ? "copro:" : "";
 const keyvPrefix = processorId + ":" + coprocessorPrefix;
@@ -56,6 +58,8 @@ const activeTasksStore_async = newKeyV(redisClient, keyvPrefix + "activeTasks");
 //   Value: data object
 const taskDataStore_async = newKeyV(redisClient, keyvPrefix + "taskData");
 
+const sharedStore_async = newKeyV(redisClient, "shared"); // Shared with Hub
+
 const tasksStore_async = newKeyV(redisClient, "tasks"); // Shared with Hub
 
 const usersStore_async = newKeyV(redisClient, "users"); // Shared with Hub
@@ -64,14 +68,15 @@ const groupsStore_async = newKeyV(redisClient, "groups"); // Shared with Hub
 
 const tasktypesStore_async = newKeyV(redisClient, "tasktypes"); // Shared with Hub
 
-
-await Promise.all([
-  cacheStore_async.clear(),
-  activeTasksStore_async.clear(),
-  taskDataStore_async.clear(),
-  //tasksStore_async.clear(), We do not clear this because it is controlled by Hub
-]);
-console.log("Cleared all KeyV");
+if (EMPTYDBS) {
+  await Promise.all([
+    cacheStore_async.clear(),
+    activeTasksStore_async.clear(),
+    taskDataStore_async.clear(),
+    //tasksStore_async.clear(), We do not clear this because it is controlled by Hub
+  ]);
+  console.log("Empty DB: cleared all KeyV");
+}
 
 export {
   cacheStore_async,
@@ -84,4 +89,5 @@ export {
   usersStore_async,
   groupsStore_async,
   tasktypesStore_async,
+  sharedStore_async,
 };
